# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "customer@prepareurself"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/get-courses": {
            "post": {
                "tags": [
                    "Courses"
                ],
                "description": "Get all courses data",
                "operationId": "App\\Http\\Controllers\\CourseController::wsGetAllCourses",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{[error_code=>0,msg=>'success']}"
                    }
                }
            }
        },
        "/api/get-all-preferences": {
            "post": {
                "tags": [
                    "User"
                ],
                "description": "Get all Preferences for a User available in Prepareurself",
                "operationId": "App\\Http\\Controllers\\CourseController::wsGetAllPreferences",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{[error_code=>0,msg=>'success']}"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key_security_example": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "api_key",
                "in": "header"
            }
        }
    }
}